test: refactor, add setup.py, add tests
I updated the README with instructions for testing and development.

I used nox to follow the pattern of other Python repos. That led to
needing to add a setup.py, since nox requires it. I used TEST_BUCKET
instead of SOURCE_BUCKET to make it clear it's for testing.

I refactored the generation logic into a generate.py file to make it
easier to test.

setup.py includes dependencies, so we can remove requirements.txt.
