+++ .trampolinerc
@@ -19,6 +19,7 @@ required_envvars+=(
 # Add env vars which are passed down into the container here.
 pass_down_envvars+=(
     "SOURCE_BUCKET"
+    "TEST_BUCKET"
 )
 
 # Prevent unintentional override on the default image.
