+++ ci/common.cfg
@@ -42,3 +42,8 @@ env_vars: {
     key: "TRAMPOLINE_DOCKERFILE"
     value: "docfx/Dockerfile"
 }
+
+env_var: {
+    key: "TEST_BUCKET"
+    value: "doc-pipeline-test"
+}
