+++ generate.sh
@@ -27,5 +27,5 @@ fi
 # Add the path where docuploader gets installed to PATH.
 export PATH=$PATH:${HOME}/.local/bin
 
-python3 -m pip install -r requirements.txt
+python3 -m pip install .
 python3 docpipeline/__main__.py build-new-docs $SOURCE_BUCKET
