--- requirements.txt
@@ -1,32 +1,0 @@
-appdirs==1.4.4
-attrs==19.3.0
-black==19.10b0
-cachetools==4.1.1
-certifi==2020.6.20
-chardet==3.0.4
-click==7.1.2
-colorlog==4.1.0
-flake8==3.8.3
-gcp-docuploader==0.1.0
-google-api-core==1.21.0
-google-auth==1.19.2
-google-cloud-core==1.3.0
-google-cloud-storage==1.29.0
-google-resumable-media==0.5.1
-googleapis-common-protos==1.52.0
-idna==2.10
-mccabe==0.6.1
-pathspec==0.8.0
-protobuf==3.12.2
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pycodestyle==2.6.0
-pyflakes==2.2.0
-pytz==2020.1
-regex==2020.7.14
-requests==2.24.0
-rsa==4.6
-six==1.15.0
-toml==0.10.1
-typed-ast==1.4.1
-urllib3==1.25.9
