+++ testdata/_toc.yaml
@@ -1,0 +1,6 @@
+﻿items:
+  - name: scheduler
+    uid: '@google-cloud/scheduler!'
+    items:
+      - name: CustomHttpPattern
+        uid: '@google-cloud/scheduler!google.api.CustomHttpPattern:class'
