+++ testdata/docs.metadata
@@ -1,0 +1,10 @@
+update_time {
+  seconds: 1594843942
+  nanos: 535529000
+}
+name: "doc-pipeline-test"
+version: "2.1.1"
+language: "python"
+distribution_name: "@googleapis/doc-pipeline"
+product_page: "https://github.com/googleapis/doc-pipeline"
+github_repository: "googleapis/doc-pipeline"
